atom-ide-ui: create pluggable analytics service for atom packages
Summary:
This creates a new provider type, nuclide-analytics, for all our atom packages to use,
which is implemented by a new internal packages, fb-analytics. This can easily be accessed
through the utility nuclide-commons-atom/analytics, which exports the same interface the nuclide-analytics
exposes.

However, the nuclide-commons-atom/analytics packages cannot be used outside the atom context, since
it depends on the atom service hub. This also includes a lint rule to ensure that the correct module
is used in atom or non-atom contexts.

For atom-ide-ui packages, the provider will be requested but will not be fulfilled, however
the nuclide-analytics implements a null provider by default so the api can be safely used in
the absense of a real provider.

Reviewed By: hansonw

Differential Revision: D5088050

fbshipit-source-id: cd150f8c705556e455b9e18f0274d7cb954d6fe7