Seperate code formatting API into explicit provider types
Summary:
There's currently 3 code formatting APIs (range, file, and on-type) and I'd like to add a fourth one (on-save) for use with Prettier and format. The intention there is that we'll only format a file on-save if it has the format directive.

Instead of having one provider type that provides all 4 APIs, I think it's better if the providers are separated into explicit types (so each provider is either a range, file, on-type, or on-save provider). We currently don't have any providers in Nuclide that provide more-than-one anyways.

This is much friendlier for Flow typing and also makes the API clearer.

Copying the docblocks here for an overview:

== Range formatters ==

These accept a range to format and return a list of edits to apply.
These will always be preferred over file formatters when a range is selected.

== File formatters ==
These always return the result of formatting the entire file.
To compensate, they can return a custom cursor position to avoid disruption.
These will be preferred over range formatters for whole-file formatting.

== onType formatters ==
These are run after every typing event in a selected editor.

== onSave formatters ==
These are run whenever a selected editor is saved.
If the global format-on-save option is enabled, then file/range formatters
will be triggered on save (even if no save formatters are provided).
Obviously, save formatters are preferred in this case.

Reviewed By: vjeux

Differential Revision: D5142385

fbshipit-source-id: f51ca4172133b3264c8a833454d4ec9fd0cf093a