Fix a potential memory leak in AtomTextEditor
Summary:
`AtomTextEditor` doesn't seem to destroy text editors when it updates.
This causes a lot of memory pressure in many frequently-displayed UIs (especially datatips) which may or may not get cleaned up.

This ensures that they're explicitly destroyed. This also updates Context View to be compatible with this change (it should let `AtomTextEditor` do the destroying).

AFAICT the GC should be able to do this as well, but it's probably a good idea in case the text editor / buffer holds on to any native code.

Reviewed By: mostafaeweda

Differential Revision: D5219477

fbshipit-source-id: f675c16b73a3c020d074ae14e3705f575e08be55