Allow feature-config to be used without a FeatureLoader
Summary:
Backpedaling a bit on this - I noticed that `hyperclick` has some custom logic to work as a standalone package (which makes sense).

And I think it'd be nice if we could use `nuclide-commons-atom/feature-config` from standalone packages without having to worry about that - it's just a matter of handling the null package name case.

As reference, I made the change in Hyperclick.

Will see if the sandcastle tests pass...

Reviewed By: matthewwithanm

Differential Revision: D5211182

fbshipit-source-id: 636fd7e83b35c0515a25a49cf22d1cd419a49fd7