Diagnostics: fix diagnostics for renamed files
Summary:
If a file is renamed (while staying open), its diagnostics stop updating!

This is because the UI only subscribes to the initial file path, which is wrong. In addition, any untitled editor that gets saved won't have diagnostics for the same reason.

This creates a `getEditorDiagnosticUpdates` that handles this correctly by listening to `onDidChangePath`.

Changelog: Fixed stale diagnostics after file renames.

Reviewed By: matthewwithanm

Differential Revision: D5278687

fbshipit-source-id: 8492528f58af21cb95fbb63d95ebfacb95956740