Strip "atom-ide" from service names
Summary:
Damien (with the agreement of the Atom team) would like to make the service names more definitive by removing the "atom-ide" prefix.

The one exception is "busy-signal", which has been taken by https://github.com/steelbrain/busy-signal/wiki/Busy-Signal-API. I'm going to leave this as "atom-ide-busy-signal" for now since it's not really part of the LSP anyway (we should probably make the APIs compatible though).

Reviewed By: wbinnssmith

Differential Revision: D5315318

fbshipit-source-id: 10345db770b8a7ad1a4f488811c4df9430e337b3