Check in a fork of shell-quote with improved comment parsing
Summary:
`shell-quote` incorrectly parses `a#b` as 'a' with a comment. Let's fix that.

`shell-quote.js` is taken directly from https://github.com/substack/node-shell-quote, with minor modifications:
- removal of the `array-*` polyfills
- fixed comment parsing, as mentioned above

I've copied the MIT license accordingly. Note that this file doesn't need to be transpiled.

Users should not use `_shell-quote.js` directly - instead `string.js` provides `shellQuote/shellParse` as Flow-typed wrappers.
Note that the underlying functions can work with operators, which are complex to support and not very useful for us. They're treated as parse errors in the simplified version.

Reviewed By: ebluestein

Differential Revision: D5354440

fbshipit-source-id: 04473f698caa8b914f320a7840f29b1a888b7acf