Remove (most of) workspace views package
Summary:
This removes most of the workspace-views package. Until everything is switched over to the Atom APIs, we still need the compat service any types so they're still there.

I also removed `destroyWhere()` from the service in favor of seansegal's `destroyItemWhere()` utility. This means the service no longer has any methods that the Atom API doesn't, which should make converting the remaining packages easier.

Note that there shouldn't be any change in functionality here since everything is already going to the Atom APIs (via the compat layer) as of 1.17.

Reviewed By: seansegal, hansonw

Differential Revision: D5449197

fbshipit-source-id: 64391d1110950dea08acccbb78d4aa7d2463d59c