Ensure code highlight markers are always destroyed
Summary:
In some situations, switching quickly between editors fails to clear the existing set of markers (leaving dangling markers around).

Make sure that we always clear the existing set of markers before creating new ones.

I also fixed the Observable a bit to be a bit more correct.
If `x` is a cursor change event and `y` is when the highlights come back, we can run into this situation:

```
x-------x-----....
\
 \---------y
```

where there's a brief period of time in which we display stale results.

Instead we should be doing

```
x-------x-----....
\       \
 \-----  \-----y
```

by moving the `switchMap` outward (so that new events always cancel the old stream).
switchMap + delay is effectively equivalent to a debounce in this scenario.

Reviewed By: captbaritone

Differential Revision: D5474082

fbshipit-source-id: 206a47b26176cfb13f836587738c3acc21d8e49b