Make busy-signal icon itself be clickable
Summary:
Waleed observed that the busy-signal icon itself should be clickable:

(1) Everything else on the status bar itself is clickable

(2) Very few people will realize they can click on the *contents* of a hover.

(3) It will be very rare for two or more things to be clickable.

This diff implements the following UX: if there are *any* clickable hover items at all, then the busy icon itself will be clickable, and if clicked then it will be like clicking on all of the clickable hover items.

Reviewed By: matthewwithanm

Differential Revision: D5477012

fbshipit-source-id: 5b80fb3912763b001f6adb22e78e301b57017cf1