Busy-signal shouldn't debounce for action-required
Summary:
If we're showing a busy-signal spinner (meaning we're waiting for the computer to do something) then it should debounce by default. That's because the computer might finish quickly and in that case we shouldn't bother the user with a spinner.

But if we're showing an action-required icon (meaning we're waiting for the user to do something) then it shouldn't debounce by default. That's because the user has to perform an action, and there's no reason to delay telling them this.

Reviewed By: hansonw

Differential Revision: D5477039

fbshipit-source-id: 3fe42722cd5b024f5b2a1b9661c010278852bf88