Busy-signal tooltip shouldn't flicker upon update
Summary:
Imagine if someone wants to update their busy-signal tooltip once a second e.g. to display the current progress number.

If the user has their mouse hovering over the busy icon, showing the tooltip message, we don't want it to flicker!

This diff achieves that. It makes it so the tooltip is only updated (flickers) when the message identities change, e.g. when a new message is added, or one removed.

But when titles change, then that's done by changing an existing HTMLElement, so it won't make the tooltip as a whole flicker.

Reviewed By: matthewwithanm

Differential Revision: D5524145

fbshipit-source-id: 1a609ad9cd2b006e717f40662358fcebcbc734f6