Deprecate diagnostics service APIs
Summary:
We're supporting a lot of APIs here and (now that Atom linter supports push diagnostics) not really getting anything out of it. LSP and Atom Linter are both standards in the community, so let's just strive to support them as best as possible.

Internally, of course, we'll still convert the callback-based APIs to Observables for easy async management and interop with other Nuclide tools.

Reviewed By: hansonw

Differential Revision: D5569240

fbshipit-source-id: 298fe552a53cbfdb2c08a895e3956bcab816edf4