Fix file tree scrollbar rendering issues at startup
Summary:
This ended up being a really tricky bug.

When custom scrollbars are enabled, a Chromium bug prevents existing scrollbars from being redrawn - so if the custom styles load after the scrollbars have already rendered, you get ugly unthemed scrollbars.

In the case of the file tree:

1. The file tree is rendered during the deserialization phase.
2. `nuclide-ui-panel-component-scroller` is styled during the activation phase. It appears that the deserialized file tree will render its scrollbar at this point, triggering the bug.
3. Theme styles load after all the packages load (which are supposed to theme the scrollbars across Atom), but at this point it's too late.

The fix is just to enable custom scrollbars by default in the scroller stylesheet when the scrollbar is visible, so when the theme loads Chrome already has custom scrollbars enabled.

I checked all the major Atom themes and they all use custom scrollbars when scrollbars are set to "alwyas show", so this should be safe.

I think this also explains why switching from trackpad -> mouse also leads to white scrollbars. In default (trackpad) mode Atom actually doesn't enable custom scrollbars... so I imagine what happens is:

1. you plug in the mouse
2. Mac OS tells Chrome to start rendering scrollbars
3. Atom is informed of the scrollbar display change
4. Atom attempts to load the custom style, but it's too late.

This doesn't solve this particular issue, since we're still too late in this case. We could listen for the scrollbar change event and force redraw all the scrollers, but I'm not sure if that's worth it.

Reviewed By: jgebhardt

Differential Revision: D5648304

fbshipit-source-id: 66e2a0b97eddc190015f109e6f30f18793fe760c