Added missing type imports for Store type
Summary: When I ran my server on all of Nuclide, it found that Store was a missing type import.  I'm not sure why there isn't a Flow error here (Is there a reason this doesn't need to be imported?) but even if there isn't an error it seems like it is worth importing the type as it allows you to hyperclick to go to the type definition + see type information on hover.

Reviewed By: hansonw

Differential Revision: D5653783

fbshipit-source-id: 1ea4186acb8adb189c47212f2d8884bc22145966