If a Diagnostic is associated with CodeActions, display them on the Diagnostics datatip.
Summary:
If a Diagnostic has CodeActions associated with it (there is a provider that has returned a CodeAction for that Diagnostic), then display them on the Diagnostic's datatip. The UI is unchanged if there are no CodeActions.

Because the UI is unchanged if there are no CodeActions, if there are no CodeAction providers, this diff will have no effect on a user's experience with diagnostics.

Reviewed By: hansonw

Differential Revision: D5626857

fbshipit-source-id: 11dca6e33ecd34654175cc985ae049e635f1ac52