Rename DiagnosticUpdater methods
Summary:
The `onDidBlah()` methods are explicitly supposed to emit immediately
upon callback registration, however this naming convention is used in
Atom for methods that *don't* do that
(`atom.workspace.onDidChangeActivePaneItem()`, etc.). Instead, methods
that invoke the callback synchronously are named `observeBlah()`
(`atom.workspace.observeActivePaneItem()`, etc.). This updates our
class to match that convention.

While I was in there, I also renamed the "all messages" methods to
remove the "all." This is in anticipation of project messages
eventually becoming special forms of file messages (e.g. with a `null`
or directory path), as is currently under consideration for LSP.

Oh, and I switch the arg order of `observeFileMessageUpdates()` to callback-last.

Reviewed By: hansonw

Differential Revision: D5690459

fbshipit-source-id: fb1becf6a6ad7c48d6179b2002f57d746d3580ca