Datatips: Hide datatips when the text buffer changes
Summary:
This is mainly for the benefit of code actions (so the actions datatip hides itself immediately).
However, I think this is a sensible thing in general as a buffer change can easily invalidate any kind of contextual information.

I took this opportunity to add the `TextBuffer.onDidChangeText` API which was introduced in Atom 1.18.
It only fires one event per transaction, which reduces the amount of thrash in most cases.

Reviewed By: matthewwithanm

Differential Revision: D5708873

fbshipit-source-id: 535f9544c5ab55c94ad5ea51c58ce7d19312782e