Move most of nuclide-test-helpers into nuclide-commons/test-helpers
Summary:
This allows us to clean up a few callsites in `modules` that rely on `nuclide-test-helpers`.

I just merged it all into one file, `test-helpers.js`. There are a few very Nuclide-specific ones that I kept in `nuclide-test-helpers`.

Note that this can't be a separate module, as this would cause a circular dependency between `nuclide-commons` and `nuclide-test-helpers`.

Reviewed By: wbinnssmith

Differential Revision: D5743590

fbshipit-source-id: 5def7bb887ed32f823bad768c5996c3d8530dae4