Create a simplified test runner for modules/
Summary:
This extracts the bare essentials from `lib/test-runner.js` and puts it into `modules`. This allows the Atom tests in `modules` to run independently of Nuclide.

Core functionality:

1) includng the transpiler
2) cleaning up temporary files on exit

The other thing from `lib/test-runner.js` that's potentially worth importing is the default config setting for `featureLoader`.. but it's easy enough to just mock it out.

Reviewed By: wbinnssmith

Differential Revision: D5751350

fbshipit-source-id: 4833926b378e74b0c9b11883dcdce573bfb9e89a