Move eslint-plugin-nuclide-internal to modules/
Summary:
This makes linting under `modules` self-sufficient.

I had to make a minor change to the `modules-dependencies` spec to have it work in atom-ide-ui and not just Nuclide.

This change actually makes it possible to remove the "rulesdir" plugin once we force developers to use Yarn workspaces (as `eslint-plugin-nuclide-internal` is going to be linked into `node_modules`!)

Reviewed By: wbinnssmith

Differential Revision: D5751346

fbshipit-source-id: 995a5d613702fc70418052bd8438e64c9119cfae