add ZIP/JAR support (1 of 4): nuclideUri archive paths handling
Summary:
This adds nuclideUri support for archive paths.  An archive path uses '!' as a separator to indicate that the preceding segment is actually a file that should be treated as an archive.  For example,

  /tmp/file.zip!abc/def.txt

...refers to a file 'abc/def.txt' contained in the archive file '/tmp/file.zip'.

Internally, nuclideUri leverages the platform path module for most path computation.  To facilitate this, we encode paths to add a path separator after every archive separator before operating on the path, and then remove the extra path separator afterwards for any return value.  For example,

  /tmp/file.zip!abc/def.txt

...would be encoded as

  /tmp/file.zip!/abc/def.txt

The platform-specific operators just consider the '!' to be part of the name of the preceding segment.  Since that is the segment we want to treat differently, this usually works out well.  This transform gives the platform-specific path operators visibility of the archive separators, and allows us to tunnel information about which segments represent archives throughout processing.

The only exception so far where this does not work out is in 'relative()', where we may not recognize a trailing segment in one of the paths as an archive.  See the comment on _matchTrailingArchive in nuclideUri for details.

Reviewed By: shushz

Differential Revision: D5795204

fbshipit-source-id: 44031874b5bcdc5e086df679c3dad2b5b27fb34d