Make addTooltip fast
Summary:
The addTooltip utility used to create a function that would *always* dispose the tooltip on a component update. Since we've started using tooltips a lot and using them in high-update throughput components it became unbearably costly.
This diff makes the addTooltip utility "remember" which tooltips are currently set, if request to dispose a tooltip comes it is not performed immediately, but rather put to a toDispose list until a timeout passes. Once it has, we now verify whether we, perhaps, have received a tooltip registration request for the same node. If so - the two requests cancel each other.

Reviewed By: mostafaeweda

Differential Revision: D5889861

fbshipit-source-id: 6d39639c75bc519ab22fe91663d1d9cb862092bf