Fixed AtomInput to deal with onBlur and onFocus properly
Summary:
`AtomInput` now debounces calls to `onBlur` and `onFocus` and does a check to make sure that the `AtomInput` is actually selected. If it is, `onFocus` is called. If it isn't `onBlur` is called. This check checks to see if `document.activeElement` is contained in this `AtomInput`.

Prior to these changes, other components using `AtomInput` had to do this on their own.

Reviewed By: mostafaeweda

Differential Revision: D5896281

fbshipit-source-id: ddf3ab94dd4f890fd91b066ac20e665ff69411a6