Support setting address offset breakpoints in Native debugger
Summary:
This diff adds the ability to right click an instruction in the disassembly view and add a breakpoint at the address. The LLDB backend will resolve the address to the nearest source line location. The breakpoints pane will show the file + line info returned by the backend.

This allows the debugger to set a breakpoint when debugging with no source, and to set a breakpoint on a specific instruction, rather than at the first instruction that corresponds to a particuluar source line.

Reviewed By: shushz

Differential Revision: D5927717

fbshipit-source-id: 5d5af9ddbfdd1c3ca8ec90d0027848619041f41f