Add signature help API + boilerplate
Summary:
This is the starting point for an API to implement `textDocument/signatureHelp` (https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textDocument_signatureHelp) from the LSP.

This API more or less matches the LSP format exactly, since I didn't see any obvious changes that we should make.

Note that this API is consumed rather than provided. The reason for this is kind of an implementation detail: it's difficult for clients to determine what the triggerCharacters are immediately, since it's provided by language servers. Instead providers can register themselves when `triggerCharacters` become known.

Reviewed By: ljw1004

Differential Revision: D5945624

fbshipit-source-id: cdc870c0da9188f2985c113609054a33548341c8