Table component: fix column sizing issue when shrinking and growing
Summary:
As raised in facebook-atom/atom-ide-ui#73 (and [in the code][1]), this min width logic was a little wonky. Because the state we stored had already taken into account the min width constraints, the component could not return to a correct state when those constraints changed.

After this commit, the constraints are applied prior to rendering instead. That way, if the constraints change (e.g. because the width of the table changes), we can just reapply them to the same state and get our new widths.

[1]: https://github.com/facebook-atom/atom-ide-ui/blob/901199814a6e4ef9a9bb0f7bb2b748425abb266c/modules/nuclide-commons-ui/Table.js#L300-L305

Reviewed By: jgebhardt

Differential Revision: D5983605

fbshipit-source-id: 5288911d9abac8b02850acf7793da6729ac8f8cb