Fix watching archive file
Summary:
When watching a directory or file inside an archive file, we actually need to watch the archive file in the filesystem.

I considered adding support for archive paths on the server side, but this way all the watches for the same archive file end up shared on the client.

Reviewed By: shushz

Differential Revision: D5994959

fbshipit-source-id: 5a37aaf19e9c45b7bcd344b1c93c05ca3b187d53