Remove scope from FileDiagnosticMessage (4/n)
Summary:
There's no such thing as a project scope, so let's remove all the callsites of the scope. (A lot of callsites still provide it; we'll clean that up as that's more of a trivial change).

Since I was here I changed the behavior of these a bit to use the 'endsWithSeparator' heuristic to ignore project-root diagnostics. (So, go-to-location won't throw an error anymore).

Reviewed By: matthewwithanm

Differential Revision: D6050026

fbshipit-source-id: 9d9aceeef7dc068e116a78d1d546484912ff50d4