Add react/no-unused-prop-types and ignores
Summary:
We used this in www and it was very helpful.

This rule is pretty good for the most part, but it does flag a
few false positives:

* When a function prop is called inside a callback (issue logged at
https://github.com/yannickcr/eslint-plugin-react/issues/1477)
* When a prop object is passed to a function, accesses within that
  function are not counted

Regardless, I have **disabled warnings for every single current
violation** for safety. For those violations that I'm confident we can
remove the props for, I marked with // TODO: remove disable

Reviewed By: jgebhardt

Differential Revision: D6060737

fbshipit-source-id: 774769a590b3973931b446062278925c3de9f52f