Add support for specification of an in-range landing position in the OutlineView
Summary:
This adds a third position to the Outline View - the landingPosition. This position is assumed to be between the start and an end, and it is the position to jump to when an item is double-clicked in the OutlineView.

The idea is that start and end declare the range that would trigger the highlight in the outline view, but the actual location may not be the start position, but rather the position where the name of the symbol is declared.

If not specified, the outline view will fall back to the startPosition, just as it was before.

Reviewed By: hansonw

Differential Revision: D6070486

fbshipit-source-id: c25ae9f1e84f9eb53680620baefa163d1cc5b016