Rename FileDiagnosticMessage to DiagnosticMessage, make FileDiagnosticMessage LS-specific (6/n)
Summary:
OK, now for what I actually wanted to do...

This renames `FileDiagnosticMessage` to just `DiagnosticMessage` in `atom-ide-diagnostics`, since there's only one kind of message now.

However, in nuclide-language-service land, we still need a separate type to ensure that it stays remotable. I've forked it as, well, `FileDiagnosticMessage` to minimize the necessary changes. We could rename to something more appropriate if we want to later.

Reviewed By: matthewwithanm

Differential Revision: D6050021

fbshipit-source-id: 99abab19f1507d04bb45e726a3d64147050bfcd6