Replace debounceTime with fastDebounce
Summary:
Due to the inefficiences found in `debounceTime`, I just went through and replaced all the single-argument `debounceTime` with `fastDebounce`.

There's a few callsites that also schedule on the animation frame, so I just preserved those.

See https://github.com/facebook-atom/atom-ide-ui/issues/93 for the difference in flame charts for a keystroke after this change.

Reviewed By: matthewwithanm

Differential Revision: D6107843

fbshipit-source-id: 0597fbcfe260ebd15f259882401ed1d7acb75b76