Select the query for easy replacement when opening
Summary:
This starts Omni2 with the query portion (everything but the prefix, here assumed to be of length 1) selected, so the user can quickly replace its text.

In practice this only occurs when a provider is triggered, and it was also the last used provider. When providers change, all the text is reset and only the new prefix is the new query.

Reviewed By: matthewwithanm

Differential Revision: D6116858

fbshipit-source-id: 13ba891835330e911c1ab1a96411cf130490f6b1