Fix size measurements on Windows
Summary:
As we saw the ResizeObserver is misbehaving on Windows. This adds a remeasuring wrapper to workaround this issue. Active on Windows only.
Interestingly enough, in my tests this does not add any additional layout costs, which means that the observed value is probably received after the layout was performed anyway.

Reviewed By: wbinnssmith

Differential Revision: D6169658

fbshipit-source-id: e4e64dd931fbb79f96e105671c8df44e2663013c