Memoize rows calculation in diagnostics table
Summary:
Currently, CodeActionsManager is updating diagnostics on every cursor change, forcing a re-render of the diagnostics table which can be verrrrry slow. While diagnostics providers probably shouldn't be updating that much, we should be handling it as best we can if they do (:

This commit memoizes the row creation which is by far the most expensive part of the re-render.

Reviewed By: hansonw

Differential Revision: D6206070

fbshipit-source-id: 6103148164fb43ac31bec9fc1a39aaab65a423a6