Improve TabbableContainer
Summary:
Instead of using atom keybindings, this change will use an observable to watch for keystrokes of Tab and ShiftTab.
This will remove the extra package nuclide-tab-focus, and put all of the logic in TabbableContainer (which I think makes it a lot cleaner).

Also, observables are called before keybinding events, so this will help with being able to make the commit form tabbable (coming soon by me hopefully).

Reviewed By: tjfryan

Differential Revision: D6360838

fbshipit-source-id: 394080eafc51786825074e4d3384fd946fe5bf6b