Use HighlightedText to match ranges in code search
Summary:
We can use the more generic component `HighlightedText`, as long as we can get a list of match ranges.

This also adds `getMatchRanges` in `nuclide-commons/string` which just does that (in a much more efficient way than the previous implementation).
The old implementation was O(N^2) due to the repeated `toLowerCase()` calls!

Reviewed By: a20012251

Differential Revision: D6456126

fbshipit-source-id: 6261d5feddd4da4e682401d11b67fd1fb6a5c9fb