Handle protected system sources
Summary:
Some folders, like /System/Library/Frameworks/Foundation.framework/Headers/Foundation.h
are not writable and, thus, cquery can't write the cache folders inside them.
So, if we detect that this is going to happen, we use a folder in /tmp to host the cache. These system files are accessed quite rarely and there's no need to have a full cache of them. They don't even have a compilation db.

Reviewed By: pelmers

Differential Revision: D6586231

fbshipit-source-id: 7ce1921fd1007df74233ad868f463b3dff1e00e5