Datatips: display *all* datatips on hover rather than just one
Summary:
I'm not certain if this is the right choice or not.. but it seems to be what people want. I could imagine this being overwhelming in some cases, though.

Although it defaults to true, I added a config 'killswitch' (`atom-ide-datatip.onlyTopDatatip`) in case some minority of users start to complain. I expect that this will be received positively, though.

Reviewed By: captbaritone

Differential Revision: D6598697

fbshipit-source-id: c23dc50bb649f22f545a3da7209b38e82e27f061