Move diagnostics gutter tip below line
Summary:
Currently the popover which appears when you hover a diagnostics icon in the gutter occludes the line to which it refers.

This diff shifts the popover down one line and to the right of the gutter.

Not only does this make the relevant line of code more visible, it also keeps the gutter icon visible which is a step toward unlocking a "click-to-persist" behavior for the diagnostics gutter icons.

This change adds some complexity to the process of moving your mouse from the gutter icon to the popover, since there is now some space between the two. To allow users to traverse this space without the popover closing, we track the user's mouse movement as a vector and leave the popover open as long as the user's mouse is consistently moving "toward" the popover.

**Before**

{F117575013}

**After**

{F117575015}

Reviewed By: matthewwithanm

Differential Revision: D6603928

fbshipit-source-id: 96b31913a5e7957a448f56ec7ab869aa0280c390