pipe stderr to a file and move the cquery_cache folder to TMPDIR
Summary:
diagnostics and stderr can grow a lot and is not good to have them next to the compilation database because they might be never removed and by default they are always appended, which means that they will grow to infinite.
Old logs are not really useful, so moving all these to TMPDIR for automated cleanup makes sense

Besides, nuclide trims stderr to 600 chars (LspLanguageService.js:414, which results in a huge lost of information useful for debugging. In order to solve that, I'm piping it to a stderr file next to diagnostics.

Also, now diagnostics and stderr are next to .cquery_cache

Reviewed By: pelmers

Differential Revision: D6714032

fbshipit-source-id: 161aa4b12b9de59bb8ad11ec0ede4d5f34840d06