Ignore package json's name and always use basename as the package name
Summary:
I'm not sure exactly when this happened, but Atom began using a package's basename to its directory as the canonical name for the package.

In our tests we passed an explicit name for the package that matched that of the package json. This no longer functions.

**In practice, we have never run into this issue since our package names often are the basename of the package anyway**.

i.e. `atom.packages.loadedPackages` maps package basenames to internal Package objects.

The "name" field in `package.json` is completely ignored in the Package object and is instead stored as "metadata" on the Package along with the rest of package.json

Instead, the FeatureLoader and its api now always take paths to packages, and never force you to provide a name, instead deriving this from this path (which is used internally for other purposes too).

Reviewed By: hansonw

Differential Revision: D6746143

fbshipit-source-id: 0b6177c04aa7c6e6729da1e5b92d74d74d87c711