Add support to showModal for onOpen and onCancel callbacks
Summary:
`onCancel` is called whenever `showModal` dismisses your modal for you in the case the user has either triggered core:cancel or clicked outside the modal. This is useful for differentiating between a dismissal from a cancel and one from a success triggering it.

Similarly, `onOpen` gets called roughly when the modal has opened.

Reviewed By: matthewwithanm

Differential Revision: D6691667

fbshipit-source-id: 0766503edac936c33dd244ef293ae8a7a3870390