Pass a node's own path when rendering it and use that to determine highlight
Summary:
Rather than have a node not know about where it exists in the tree, and having it told whether or not it is highlighted, pass each node its own path and determing highlight status within the node.

This will allow for easily passing a node's path up the tree when it's selected.

Reviewed By: matthewwithanm

Differential Revision: D6775828

fbshipit-source-id: 47a252d7e42da13d22050ca22ffe7f2e0a606b4a