Add collapsability
Summary:
This adds support for collapsability to SelectableTrees.

Like the list of selectedPaths and appropriate props to handle selections, this introduces `collapsedPaths` and `onCollapse` and `onExpand` handlers.

It also involves a fairly gross measurement of the `:before` pseudo element and its parent node to determine if the user clicked on the disclosure triangle to the left of an outline node. Ideally this is replaced with hiding the theme's pseudoelement and implementing the chevron triangle as a real node.

Reviewed By: captbaritone

Differential Revision: D6794819

fbshipit-source-id: b295d1442c33fa94a341a401e879bf58c54d310e