Extract previewOpenable from Dash
Summary:
Extracted from Dash's openable previews, this opens a pending pane (if the user has them enabled) to the requested uri. It also accepts options for GoToLocation such as row and column.

If a row is provided, it is highlighted with a marker. If both a row and column are provided, higlight the row with a marker and apply a point marker to the row + column point.

Reviewed By: matthewwithanm

Differential Revision: D6841730

fbshipit-source-id: 33ec7ac198a8dca7d38a91cd723412f5fe3db2bc