Implement a custom outline tree parser
Summary:
Beautiful child of mine!

This parses the odd but not incorrect outline data from cquery and tokenizes it so that it has some colors. Not a lot of colors, because that would be confusing.

It supports pretty much every single symbol in c++ and obj-c. It doesn't support typedefs, but pretty much anything else if supported. There are even special handlers for operators.

It looks beautiful and full of colors

Reviewed By: pelmers

Differential Revision: D6814114

fbshipit-source-id: dc3acc43539086c9d9b81d81c9e190324df3f162