Complete process observable regardless of input observable
Summary:
If you provide a `input: Observable<string>` to `observeProcess`, it will not complete the process until the input observable is completed. If the process exits before this happens, then you will have a lingering process observable that just hasn't `completed` yet. Instead, we should only take the input observable for as long as the process is actually running. This should be fine since any input would be disregarded (since the process has exited), and if a completed observable tries to emit values nothing happens.

Presumably no one is trying to send input to a closed process on purpose, and hopefully no one is using an input observable to keep around a process observable they don't want to `complete` even after it has exited.

Reviewed By: matthewwithanm

Differential Revision: D6992810

fbshipit-source-id: 06e7300542a29a8de902921262335ea0e1870e8d