Fuzzy File Provider
Summary:
This adds a fuzzy file provider to Dash. Powered by the same service that powers the existing provider in QuickOpen.

Currently uses the prefix 'f', but I think we can do better.

Leverages showing file icons in OpenableResults in a prior diff.

Leverages previewing files behind the modal which landed a few weeks ago. But for me this is quite slow, and I'd like feedback from the UI team on this.

This only ships to the Nuclide UI team.

Reviewed By: captbaritone

Differential Revision: D6967178

fbshipit-source-id: 1082cc98c16a26535670fb2b4b02e62320992b55