Fix editor.getNonWordCharacters() callers
Summary:
In 1.24 the `editor.getNonWordCharacters` (undocumented) API changed to take a position rather than a scope name. This creates a wrapper, since we'll still have a lot of 1.23 users in the near future.

(It looks like, going forward, that this API is no longer going to be super useful.. in tree-sitter grammars this just returns a constant. It doesn't seem to fit the tree-sitter model that well.)

Reviewed By: ljw1004

Differential Revision: D7112975

fbshipit-source-id: 4df6b8cc4dc6f31b829ee8756e18c5df70ae30f0