switch package states from binary (T/F) to enum
Summary:
Previously, nuclide features were saved in the config in one of two states: `true` (enabled), or false `disabled`.

The problem with this is that there are some packages that are off by default, and we want to differentiate between a package being explicitly toggled by the user, and the package just being in its default state.

This diff introduces a new system for keeping track of config states (a ternary enum), where 'always' or 'never' means that the enum was explicitly toggled by the user, while 'default' means that the package is in its default state (and therefore has not been touched by the user, or it was touched and then returned to its default state).

Reviewed By: matthewwithanm

Differential Revision: D7057182

fbshipit-source-id: 4b79e76dcdd6b977adae48aba4a665f7d1ae876c