Restore support for icons passed directly.
This was inadvertantly removed when adding support for titles for
outline nodes that gave a `kind`.

In the case the passed icon is itself an Atomicon, we look up the
corresponding name of the Atomicon and use that.

This also removes the `aria-label` set on the icon as the title should
suffice.