Nuclide debugger: add experimental support for TerminateThread
Summary:
This diff adds support to the UX for TerminateThread, which is currently proposed as an addition to the VS Code debug adapter protocol, but not yet added.  A debugger would indicate support for this in capabilities, and if supported, Terminate Thread would appear as an option in the threads pane, and on the threads pane context menu.

We currently send the terminate thread request with an "fb_" prefix since this is experimental. The prefix would be removed once this command is added to the protocol officially. Our HHVM debugger will support the "fb_" prefix.

Reviewed By: mostafaeweda

Differential Revision: D7198520

fbshipit-source-id: 9189646901223cdeb6648a48833f31a1303e460f