Add AbortController/AbortSignal polyfills and utilities
Summary:
This adds polyfills for `AbortController` and `AbortSignal` based on the spec here: https://dom.spec.whatwg.org/#aborting-ongoing-activities

These have landed in Chrome but won't be out until Chrome 66, so we'll be stuck with these for quite a while. (It's also unclear what the story is when it comes to Node integration).

(NOTE: there's https://github.com/mo/abortcontroller-polyfill but its implementation seems a little not-up-to-spec.)

I used polyfills for `DOMException` and `EventTarget` as those are not available in Node (even in Chrome, `EventTarget` isn't instantiable until Chrome 64.)

I also added two utilities for interop between RxJS and `AbortSignal` in `observable.js`:

- `fromAbortablePromise`, which takes a function that accepts a `AbortSignal` and returns a `Promise<T>` and turns that into an `Observable<T>`. Unsubscribing from the observable naturally sends an `abort()` to the `AbortSignal`.
  - This is meant to be used to e.g. convert cancellable fetch to an Observable.
- `toAbortablePromise`, which turns an `Observable<T>` + `AbortSignal` into a `Promise<T>` which throws an "AbortError" `DOMException` upon receiving an abort.
  - This is meant to help whken you want to actually implement a cancellable function.
  - It takes a nullable `signal` to work well with the spec for cancellable functions, which indicates that `AbortSignal` objects should be accepted through a `signal` dictionary member (which may not exist).

Reviewed By: matthewwithanm

Differential Revision: D7314270

fbshipit-source-id: 2818fe81d62029fb2cc9de3a940d48fec72ed227