Add rootPath option for LSPs in .bigdig.toml.
Summary:
When we launch the Reason/OCaml LSP in www, we need to ensure that the
rootUri/rootPath is `www/scripts/re` rather than just `www`. Previously, we
always used the root of the remote directory that you mounted, which is fine
for Hack and Flow.

Admittedly, things did work if you mounted `www/scripts/re` as the root folder
in VS Code, but that was somewhat annoying as then you couldn't see any other
files in www if you wanted to. This gives the best of both worlds.

This ended up becoming a ball of yarn that is a bit difficult to untangle.
Rather than continue to rely on `jsonrpc.StreamMessageReader` to intercept
the `initialize` LSP method and rewrite message, we upgrade the version
of the `vscode-languageclient` module to take advantage of the new
`registerFeature()` method that can take a `StaticFeature` which has a
`fillInitializeParams()` method where you can modify the `initialize` method
without paying the extra cost of parsing the stream an extra time (though
at the moment, we still do because of the new logging facility I introduced
in D7418590.)

By eliminating the need for `jsonrpc.StreamMessageReader`, I was able to
simplify how we proxy the stdin of the LSP process. (Though again, we still
have to make an extra copy of the stream contents for a side channel that
we need in the event that logging is enabled. There is a TODO about using
a "Middleware" to eliminate this cost.)

In order to get the new version of `vscode-languageclient` to typecheck,
we needed to introduce Flow definitions for both `vscode-languageclient`
and `vscode-languageserver-protocol` (because that is where `InitializeParams`
and some other types used in `StaticFeature` are defined). Because I used
the code in this revision to drive the new Flow definitions, it's now a bit
difficult to split these changes out. Introducing the Flow definitions on their
own appears to make the repo not Flow clean and at first glance, it seemed
like much of the work I do in this revision is required to clean it up.

Reviewed By: hansonw

Differential Revision: D7423989

fbshipit-source-id: 87c4a50e831ae9e46e891ad4b36345ed12212043