Prevent signature help's datatips from adjusting to the editor height
Summary:
While this isn't ideal, signature help's datatips use CSS to adjust their position to be above the line (assuming that Atom puts the marker below the line).

However, Atom can adjust the position of the overlay to better fit within the window (which breaks the positioning). Ideally we'd have a better solution for this but for now the easiest fix is to disable the automatic position adjustment for `above-range` pinned datatips.

Reviewed By: pelmers

Differential Revision: D7478238

fbshipit-source-id: b6db5bd85ad50c696415d30f8271b2c665b7ca1d