Fix console height when filter UI is present
Summary:
Previously, when you added a source filter, a filter UI would be added to the top of the console, but the height would not be remeasured. This made it impossible to actually scroll to the bottom of the console.

This also removes another instance of ResizeSensitiveContainer which is deprecated.

Reviewed By: wbinnssmith

Differential Revision: D7488876

fbshipit-source-id: fad0c0f4503d3c42fdff8be2bf0d702e3634ea16