Investigate bad callers of `goToLocation`
Summary: I found the error `editor.scrollToBufferPosition is not a function` when checking our logs. It is happening because `atom.workspace.open` returns `null` in some situations. I got hundred errors when I change the type signature. Some callers don't bother about the return values and others do. This log will help me understand if the problem happens in some special case or it is a general issue.

Reviewed By: hansonw

Differential Revision: D7624697

fbshipit-source-id: 9ebbfaeebafbdf20b1fa6428cc3a960bfe9a1a61