jest-atom runner
Summary:
sorry for a huge diff. it was a spontaneous hack :(

this implements a jest runner that can spawn atom processes and run tests in them.

Video:
https://pxl.cl/d2bF

to run tests you need to run jest and pass the config for atom tests:
- `cd modules/jest-atom-runner/`
- `yarn watch`
- `jest --config ~/nuclide/jest/atom.jest.config.js`

This will pick up any tests that are located under any of the `__atom_tests__` directories within `nuclide/` project.

This is not run on any of the CIs/Sandcastle yet.

next step would be:
1. dogfooding it and making sure it works reliably
2. adding a sandcastle step to our macos build
3. polishing (the IPC part is gnarly :( )

Reviewed By: hansonw

Differential Revision: D7771785

fbshipit-source-id: 68dbce1d95d0f9e90a68718b43b4523489d3973c