Atom IDE Debugger: Updated API for Setting Debugger Configuration Providers
Summary:
The problem with the old API was that it wasn't safe because each provider was called regardless of if it cared about the current configuration. This new API uses the adapterType to only call resolveConfiguration if the provider deals with that type of adapterType.

Additionally I have added a UniversalDisposable to the IProcessConfig so that the resolveConfiguration can also add disposables that trigger on session end.

Reviewed By: ebluestein

Differential Revision: D7836534

fbshipit-source-id: 9b4d3fff6b40b6dbeb4f3ea2e87b2c2abf0796f1