Add auto adb tunnel feature to devservers
Summary:
I added this mostly to speed up my own development. Not sure yet if we want to roll it out more broadly in the future.

This basically creates a new adb tunnel for every new project on a host we haven't seen before and shuts it down when there are no more projects opened for that host.

Reviewed By: shunfu

Differential Revision: D7889063

fbshipit-source-id: e666bb7a0ce647bc06c29681b9df8a90179c59c2