Get Flow to understand isValidTextEditor
Summary:
Flow supports the "%checks" annotation which allows a function to extend its type refinements back to the caller.

Ideally we'd just modify `atom.workspace.isTextEditor` to have this property but %checks isn't supported for methods :(

I grepped for the callsites and cleaned them up as best as I could. Some places were doing `observePaneItems` and then filtering which is basically equivalent to `observeEditors` :P

Reviewed By: drmathochist

Differential Revision: D7885798

fbshipit-source-id: cefc0224f9fd10fb937fde7658e34f1f193bbd15