Enable "@flow strict" and "@flow strict-local" in Nuclide codebase
Summary:
Flow strict is a mode where several previously-allowed patterns trigger a Flow error. This mode is configured by creating a section `[strict]` and listing out the lint rules that should run in this mode. Flow docs [suggests](https://flow.org/en/docs/strict) to enable the following rules:
- nonstrict-import
- unclear-type
- untyped-import
- untyped-type-import
- unsafe-getters-setters
- sketchy-null

This diff implements the ground work for the adoption of `flow strict(-local)` in Nuclide by
- adding strict section in Nuclide's flow config with the suggested lint rules from the Flow docs
- supporting `flow strict` and `flow strict-local` in our `license-header` lint rule.

**What is the difference between `flow strict` and `flow strict-local`?**
The later doesn't enable `nostrict-import`. So, we can move files to a more strict env without having to make sure that its dependencies are strict. Eventually, we will be able to flip the switch and make more files `strict`.

**What is the benefit of this change?**
This change won't improve the type coverage but it will helps to not degrade our typing coverage since we won't be able to use `any | Object ..` in files with `flow strict(-local)`. I can see that adopting strict mode makes the prototyping hard because we can't play safe but it will pay back when we have a nice night of sleep after a day of refactoring sat on flow's shoulder.

**How many nuclide files will be strict?**
We are going to add `flow strict` to [~337 files](P59509464) and `flow strict-local` to [~1013 files](P59509458) from a total of [~3542 files](P59509907).

**How much it is different from what we have today?**
Nuclide's flow config enable `untyped-type-import` and `sketchy-null` lint rules as error. It means that we are going to notice only `untyped-import` and `unclear-type`.

**What is the impact on flow migration?**
That is a good question that I don't know how to answer. If it starts to add to much work for the oncall, we can always remove `strict(-local)` to the files.

**Next steps and opportunities**
- Use `flow strict-local` in our JS templates
- [Run a script](P59509527) to increase the adoption
- Add flow definition for third-party libraries (e.g. `nuclide-uri` ..)

Reviewed By: semmypurewal

Differential Revision: D7856708

fbshipit-source-id: 472430d3cfccf89365b27208000b6314ca443065