Ensure that nuclide-node-transpiler is a devDependency of all modules
Summary:
nuclide-node-transpiler shouldn't be included in production. This rewrites `atom-ide-debugger-native-gdb`, `atom-ide-debugger-ocaml`, and `nuclide-debugger-cli` to avoid the requirement by adding a `DEVELOPMENT` file that may be checked (this is the same approach used in the `big-dig` packages.)

`nuclide-debugger-cli` doesn't need this because it's only used in the fb-only code, which can just assume that the code lives inside of Nuclide.

(I cleaned up `fb-nuclide-start-cli` to avoid some duplication.)

Reviewed By: ebluestein

Differential Revision: D7915370

fbshipit-source-id: 639a5ea092d1413f9faa8092730dfea7bb8a1ebe