Stop using eslint-plugin-rulesdir for nuclide-internal lints
Summary:
The `rulesdir` hack was necessary when ESLint was upgraded, because there was no way to import rules outside of node_modules without it.

Luckily with Yarn workspaces `eslint-plugin-nuclide-internal` gets symlinked to `node_modules` by Yarn... so we can go back to using the more descriptive `nuclide-internal`. Yay!

I used the codemod tool to find+replace `rulesdir/` throughout the codebase.

Reviewed By: captbaritone

Differential Revision: D7940078

fbshipit-source-id: 12280e08b4ff7adc59fd0bd94c346b17d0e90e64