Defer feature loading
Summary:
This defers feature loading until a project is opened. (Project = a project file or ad-hoc one created by adding a folder.) The result is about 20% faster activation. Of course, this is only when you're opening Atom without a folder and you pay a cost [1] when you add your first folder. I think this is a reasonable tradeoff, though…the hope is that project files become more prevalent and then you never have to pay that full cost (since you only need to load a subset of features).

Here are some measurements I took of the timeframe starting at the beginning of `FeatureLoader.load()` and ending after the final feature was activated. Note that this is just measuring the synchronous activation time and doesn't account for the cost of scheduled work or include the benefits of not having to deactivate features on project load, etc. These are the average of 10 runs. I hid the file tree and loaded a very small local folder to try to eliminate that as a source of delay.

Without deferring:
Nothing Mounted (Baseline): 2958ms
Folder Mounted: 3035ms (+2.6% — we would expect this to be +0%)
With deferring:
Nothing Mounted: 2346ms (-20.7%)
Folder Mounted: 2934ms (+0.8% — we would expect this to be +0%)

Previous measurements were even better, but I was missing a bunch of atom-ide-ui packages in the nuclide-core group 😞. Also, I tried with a new "nuclide-initial" feature group that was MUCH more minimal and saw an order of magnitude faster startup, but that was at the cost of UI weirdness (since a lot of UI-adding packages were loaded later). We should investigate this finer-grain control later, though, as this can drastically improve the startup time.

[1] Something interesting is that the cost you pay when adding the folder later *isn't* the full difference (it's much less). So we may just be benefitting from spreading the work out a bit (instead of throwing everything at it at once), but I haven't investigated further.

Reviewed By: hansonw

Differential Revision: D7824569

fbshipit-source-id: 4eecada79b851438612d055aeea591f92f0f71df