Make length of "starting debugger" more accurate
Summary: Previously we transitioned from the STARTING state to the RUNNING state right after launch or attach. This meant that for attach or launch sequences that took a long time, we switched to "Running" way too quickly. Move that transition to after ConfigurationDone, where some debuggers do a lot of work.

Reviewed By: ebluestein

Differential Revision: D8049692

fbshipit-source-id: e7b941104521da1c82b9696ffb25c3d9cca85c63