Handle async Electron menus
Summary:
In Electron 2.0+, all menus created via the `remote.Menu` API are going to be asynchronous (i.e. non-blocking).

This requires some changes in the code - in particular, when used with React, we should tie the menu lifecycle to the component lifecycle. (Otherwise you could end up with a dangling menu attached to an unmounted dropdown, which could be bad.) I enabled the `async` flag to be consistent, and so we can test for any issues with Atom 1.27.

This also has the nice side effect of preventing menus from triggering stalls (previously they were indistinguishable!)

Reviewed By: wbinnssmith

Differential Revision: D8164875

fbshipit-source-id: e45c9a3aed6bc518e76bbc7ff2f608d96a5ced9a