Prepack Debugger: Fix Regression
Summary: There was a regression where the debugger stopped working because of an incorrect path. It was broken because it was looking for the debugger in `nuclide-debugger-common` instead of `nuclide-debugger-vsps`. Additionally after fixing that, it is clear that the debugger does not look for a default path for the cli so I have updated the message to not say that it uses a default path.

Reviewed By: mostafaeweda

Differential Revision: D8174009

fbshipit-source-id: 28bf669afe0ab490feb63be7b85dbd0655e33d4e