Ensure user-entered source root is correct
Summary: We are now filling in the computed source path for the proper buck root if the user doesn't enter one; however, if they do, we need to make sure it's the canonical path. Run realpath against whatever they enter to ensure it matches up against what the debug stack will produce.

Reviewed By: ebluestein

Differential Revision: D8359137

fbshipit-source-id: d92fa0a056b1ecee53c6a9fc3d27fd3f06e17636