Remove some dependencies on atom's Directory class
Summary: We were creating and passing around Directory objects in a lot of places but ultimately we only used them for functionality we could get with strings and nuclideUri utils.  'pathwatcher' was introduced as a way to use Directory without direct atom dependency, but led to build errors, so that has also been removed.

Reviewed By: hansonw

Differential Revision: D8324414

fbshipit-source-id: 9f8e8482126ad56261e1b8daa25203c2642d004b