Isolate sdb/Tizen code from nuclide-adb
Summary:
Stop all Tizen code from depending on nuclide-adb, so we can clean up certain abstractions serving both.

This diff seems massive, but most of the Tizen code is on life support and I'm not interested in maintaining it going forward. Most of the additions are simply duplicated, so they keep working, and they can be easily deleted in the future without breaking anything else.

 ---

One thing that I should point out is that this removes the `Configure adb` button from device panel.

Reviewed By: aadsm

Differential Revision: D8410929

fbshipit-source-id: 8377521a02bb4ef87b4a5a43d188890d090546dc