Replace manual onDidDestroy tracking with UniversalDisposable.addUntilDestroyed
Summary:
I grepped for "editor.onDidDestroy" and looked through a bunch of them. A lot of them are trivially replaceable with `addUntilDestroyed`!

Sometimes you end up with

```
disposable.addUntilDestroyed(editor, editor, ...)
```

which looks kind of weird but makes pretty good sense if you think about it - we want to destroy the editor when the package exits, but not if the editor is destroyed first.

It also makes sense to use

```
disposable.addUntilDestroyed(editor, editor.onDidDestroy(() => { ... }))
```

which prevents the `onDidDestroy` subscription from leaking. (Though `onDidDestroy` already uses `once` internally so it's already not a major leak.)

This fixes at least two potential memory leaks: one in `BreakpointDisplayController` and another in `nuclide-commons-atom/text-event`.

Reviewed By: matthewwithanm

Differential Revision: D8372728

fbshipit-source-id: 2020ebb8cd9e94e3a8884862bda91f79396acf6c