Bugfix: Android Debugger Works Even if SDK Sources Are Unavailable
Summary:
Before this diff, if SDK Sources were unavailable, the android debugger would fall over and crash. Now it shows a warning and proceeds to debug anyways.

I also added tracking so that if we see that there are large numbers of people without the sdk source path, then we might have an issue (likely, we don't have the correct ANDROID_SDK env variable or the correct path to the sdk manager).

Reviewed By: ebluestein

Differential Revision: D8651891

fbshipit-source-id: 9fe67f2d4ae0d02fc7d056fc87f99e90bc21d54b