Use IndexedDB for RecentFilesService
Summary:
Changed RecentFilesService to use IndexedDB instead of Atom serialization.  Used the `idb-keyval` kv-store because we are just using IndexedDB as a stable store to back a LRU cache of recently-touched files.

It seems like we should be OK if we quit in the middle of a transaction (i.e. on final Nuclide quit if the save doesn't happen before the window is closed).  Some quick searching found that it's [implementation-specific]( http://lists.w3.org/Archives/Public/public-webapps/2012JanMar/0944.html
) how that's handled, but there are no real issues around it [unless you are out of disk space on Windows](https://bugs.chromium.org/p/chromium/issues/detail?id=146284).  There are specs about [aborting a transaction](https://www.w3.org/TR/IndexedDB/#transaction-abort) and it seems like that they try to abort a transaction if it's interrupted.

Reviewed By: wbinnssmith

Differential Revision: D8640543

fbshipit-source-id: 2eadb43deb38c337554b94ed18f151a556becfc2