ThreadTreeNode will still show arrow even if there are no child elements (stack frames)
Summary:
A "UI bug" exists: even if there are stack frames for a thread, the ThreadTreeNode may not have the drop down arrow next to it. This diff will always have ThreadTreeNode have a drop-down arrow next to it.

Unfortunately, ThreadTreeNode will have a drop-down arrow next to it even if there are no stack frames. However, that seems to be the standard for Nuclide, in particular, the File Tree. If you open an empty folder, the drop-down arrow still displays, and you can still expand/collapse the empty folder.

Reviewed By: mostafaeweda

Differential Revision: D8725549

fbshipit-source-id: 86ce8a5b6df079ed918eec4c960f97dc83014c1f