Switch atom-ide-rename keybinding to `cmd-r cmd-r`
Summary:
Previously, the `F2` keybinding was not friendly towards Mac Pro users with touchbars. But the sequence-keystroke `Cmd+R, Cmd+R` seems to be a more comfortable and mnemonic solution to this.

This can be a pattern that we follow when keybinding future refactor features as well.

One thing, though, is that the context menu still only shows `CMD-R` as the keybinding for `Rename` and when clicked, does actually immediately open the UI. I'm not sure how to change the display of the keybinding on the context menu at the moment but I'll have to figure out how to fix this in the future.

Reviewed By: mostafaeweda

Differential Revision: D8740576

fbshipit-source-id: 40ecee185580058375bbef23c26f9cd49eac13b6