Change Dash Description to use KeyBindings
Summary: Use kbd and determing KeyBindings for the Dash Onboarding task. Add a `findKeyBindingsForCommand` function in nuclide-commons-atom (TODO in a later diff: use this function wherever keybindings for atom commands are obtained) which determines the keybindings for an atom command on the user's platform. Use this function and pass in the obtained key bindings for the Dash onboarding task kbd element in the description. Use an Observable for the Dash command to pass down the generic or true omni command.

Reviewed By: wbinnssmith

Differential Revision: D8764714

fbshipit-source-id: da648598780545baccb90902e589988aeea88f2b