Change ocaml projectRoot search strategy
Summary: We had a sketchy algorithm for determining ocaml project root. It was for instance deciding that "/home" was a good project root, because there exists a user named esy and so "/home/esy" exists, and this looked like it might be a esy project root. The comments in the code give a detailed explanation about what the algorithm should be for picking project root, and why.

Reviewed By: mostafaeweda

Differential Revision: D8832248

fbshipit-source-id: 64b78dd4d81855b0b170c16b34e4624144fef239