DiffPreviewUI integrated with inline-rename
Summary:
Local renames will be immediately executed, but when the rename entails a change across multiple files, a preview of the files to be changed and their specific lines will be shown to the user for confirmation. Upon doing so, the changes will be written directly to disk.

In the former case of local renames, it is assumed that the renames will be on a currently open file, so the changes will be applied on to the buffer (i.e. file won't be automatically saved). Otherwise, executing a local rename on a closed file will stop the operation and throw an error message.

Reviewed By: mostafaeweda

Differential Revision: D8822721

fbshipit-source-id: 9ccadcb6d039b2c2729e31b4175f2dee7c96ea16