Only sync and fit when visible
Summary:
Noticed a strange interaction when switching tabs back and forth with the DOM renderer: upon switching back often the terminal would be cleared until an interaction such as keyboard input was performed.

Xterm.js documentation notes that the `fit` addon only works when the terminal is visible, so avoid calling it if terminal is not visible (e.g. switching tabs causes the resize observable to emit before terminal is visible).

Reviewed By: kcaze

Differential Revision: D8813190

fbshipit-source-id: 0b5030b7cbd34d1c9707ef53a89c56063627fa4b