Merge device actions and device tasks
Summary:
This merges the concept of device tasks and device actions. Buttons for them are now rendered both in the device list and on the detailed device page.

They are also consumed as an observable, so a provider can send updated tasks for each device on the fly (being able to push, vs previously waiting for a poll by device panel).

Reviewed By: mostafaeweda

Differential Revision: D8825159

fbshipit-source-id: 7890dbeaa75ed7f9f1426e820e0eceb9c0ab3554