Clean up debugger menu
Summary:
Currently the items in the debugger menu are sorted randomly, making it tough to find what you're looking for. A couple items in the menu shouldn't actually be in a top-level menu and are really commands intended for use elsewhere in the UI.

This diff groups Debugger menu items into logical categories (starting the debugger, stepping the target, manipulating breakpoints, managing the debugger pane, utilities). Within each category, items are sorted by name so they're easier to find.

Finally, adding accelerator key hints. These are important for accessibiltiy and keyboard navigation on Windows.

Reviewed By: Goom11

Differential Revision: D8877909

fbshipit-source-id: c4a9136b0ca1a4a5c6265066131c76da4085525e