Cleanup call stack fetching
Summary:
This diff removes the need for the debugger UI to explicitly invoke the fetchCallStack logic to get a call stack, which allowed me to greatly simplify the new Process, Threads and Call Frames components for the debugger tree. Instead, callers get a getFullCallStack() method on IThread which returns an Observable<Expected<IStackFrame[]>>. The model will automatically fetch the call stack if needed under the hood, and the returned observable will either emit an Expected.Pending followed by the call stack if it needed to be fetched, or just the call stack if it was already available.

Fixed a few bugs along the way:
1. We were not always respecting the debugger capabilities supportsDelayedStackTraceLoading - this was causing an exception in the Java debugger.
2. There were some bugs in the expand logic of the new ThreadsNode that could cause us to expand without properly refetching the stack.
3. There was a race in the _tryToAutoFocusStackFrame logic in DebugService - when a thread stops and becomes focused, we fetch the first frame, and then queue a delayed call to fetch the remaining frames. If the first frame does not have source (but a lower frame does) we'd correctly select the first frame with source only if the second call happens to return before _tryToAutoFocusStackFrame is invoked.
4. ThreadTreeNode was fetching frames for collapsed nodes.
5. HHVM Debugger was not behaving properly if "levels" on stack trace command was 0 - it should return all frames, not no frames.

Reviewed By: mostafaeweda

Differential Revision: D8925566

fbshipit-source-id: b14c502f6c3ced01df1870d89a1c70a1e104ab7b