Don't fetch callstacks for collapsed tree nodes
Summary: This diff fixes a bug where we were fetching the entire callstack for a thread on callstack changed notification from the DebugService. We actually only need the first frame, and can fetch the callstack only for thread nodes which are expanded. This optimization reduces chatter with the debug adapter, and makes source stepping faster.

Reviewed By: Goom11

Differential Revision: D8941447

fbshipit-source-id: 4fcee5cb53f0b69590acd09c39708c41bf0da187