Simplify and cleanup debugger focus logic
Summary:
This diff cleans up the set focus logic to ensure that we always have a consistent focus state - it should be impossible to have a stack frame from one session focused without a corresponding thread and process from the same session.  Added observable events for focus changes that the UI can subscribe to.

Simplified and centralized logic for choosing what thread and process to focus on as processes, threads and stack frames come and go.

Reviewed By: kevinatorchen

Differential Revision: D8979919

fbshipit-source-id: 1255dddf6ed1b25502c810f38fda6a342a420ea6