Show "Starting" in debugger process tree
Summary: Some debug adapters take a few moments to spin up. When debugging multiple processes it's not immediately obvious from looking at the process tree that one is still starting because the message beneath the debugger stepping controls is far away.  Adding "(starting)" to the tree node makes it clear why there are no threads and the target can't be paused yet.

Reviewed By: kevinatorchen

Differential Revision: D9004057

fbshipit-source-id: 40027035fe01a60a898107a1973b04cb1721d6bd