Ensure the Dash Onboarding Task's Pane Has Focus on Mount
Summary: There was an issue where continuing to the Dash Onboarding task from the previous (Key bindings) task by clicking the continue button sets focus to the continue button, resulting in the "Pane: Split Left" command not showing up in Dash. This diff fixes that issue by focusing the active pane when the Dash Onboarding task is mounted so that Dash/Command Palette has the pane's context for commands.

Reviewed By: wbinnssmith

Differential Revision: D9020505

fbshipit-source-id: daece380b3fabda2bd1d4b7faef516cc87a9e369