Don't create a tunnel for exopackage in the adbmux case
Summary:
It looks like adbmux now handles exopackage so we don't need to tunnel this port. The worse situation is that since we open the tunnel adbmux is not able to work properly with exopackages.

I found a bug in adbmux while testing this, it seems that we're not able to dispose of a forwarding port, because of this when buck installs it doesn't close the forwarding making it impossible to create the tunnel if someone wants to buck install to a one world device and then to a real device. Raj is working on this issue: t32030543.

Differential Revision: D8995780

fbshipit-source-id: c36e85b856710b0f637bc8149482c54fd94ceb0d