Add a "conditional nux" experience based on membership in a manually curated GK
Summary:
This implements a tooltip nux to be shown to users who belong to the `nuclide_diagnostics_nux` GK.

The nux triggers the diagnostics table to open, as well as a tooltip introducing them to the table to be shown.

After the user has experienced the nux, it will never show to them again (we flag the state in indexeddb).

Reviewed By: matthewwithanm

Differential Revision: D9036244

fbshipit-source-id: 313716194fe5500ab1d375b89307d07c8f454ded