Allow atom://nuclide-terminal-view to be opened
Summary:
Previously this was impossible. The minimum url for opening a terminal was `'atom://nuclide-terminal-view/?command='` as `query` was {} otherwise. This would cause it to always take the `else` branch and attempt to parse `undefined` as JSON.

This also adds a concept of a fallback cwd, determined outside the TerminalView at toggle/open time and used if the uri does not specify one.

Reviewed By: matthewwithanm

Differential Revision: D8844266

fbshipit-source-id: aa887e075b4158676ee30fe48c0f427e80421a3b