Handling large refactorings
Summary:
Before, Nuclide would crash when trying to show a preview of all the files affected by a rename of `nullthrows` - which occurred in over 78,000 files in `www`.

Now, we truncate the listed results in the UI at 1000. Doing so allows us to render the next UI phase and apply all the edits as usual.

Reviewed By: arxanas

Differential Revision: D9111899

fbshipit-source-id: 89c6941d127bce0e39b82f876392e2c16414cadb