Handle Migration of 'fb-nuclide-onboarding.showOnboarding' to 'nuclide-welcome-page.hiddenTopics'
Summary:
Since Onboarding will be integrated with Welcome Pages, the show/hide functionality of Welcome Pages can be leveraged. To ensure users who already chose to hide Onboarding don't see it again with the integration, the showOnboarding config value needs to be migrated so that the fb-onboarding topic is stored as a nuclide-welcome-page hiddenTopic.

This diff adds a function for that migration and removes the showOnboarding config value from the fb-nuclide-onboarding package

Reviewed By: wbinnssmith

Differential Revision: D8944890

fbshipit-source-id: 74c4b780aca047a3d3785515da568367b59e4aa8