Implement Nuclide Config
Summary: Implement the nuclide-config module using atom's atom.config construction as a base point. The config is constructed using the atom$Config constructor, and its values are initialized from disk if the nuclide-config.cson file exists. The saveCallback is the same as the atom.config saveCallback except with nuclide-config context.

Reviewed By: wbinnssmith

Differential Revision: D9091031

fbshipit-source-id: 340c82b3684ea1115436a0bb0fa0e853fb37bc55