Fix global Store type usage
Summary:
The react-redux flow types are leaking a global `Store` type. Worse still, it's an `any` alias. That means that, if we reference a `Store` type that we forget to import, it's just `any` typed and we miss errors. This resulted in real file tree bugs.

I fixed the `import` in the file to not leak and then fixed the errors. The bad news is that I can't keep the typedef fixed since then we'd need to pull in the redux typedefs which surfaces a TON (~2000) of errors.

If anybody has ideas for a path forward on that, I'm all ears.

Reviewed By: zertosh

Differential Revision: D9256718

fbshipit-source-id: 05a4955385a92c8c4f462d023589c93bcb7225e8