Change the default location of terminals to be the bottom (again)
Summary: Now that we leverage a standard uri to get Atom's docks to open new terminals in the last known location of the terminal, we can safely change the default to be the bottom. If a user has decided that they prefer the pane location (the main text editor area), they can simply relocate one terminal there and all future ones will open there.

Reviewed By: pelmers

Differential Revision: D8878282

fbshipit-source-id: 787ae8dcfe5a14183c7f4913c90a0b6d3085e377