Add `atom-ide-terminal:toggle`
Summary:
Like console, diagnostics, and pretty much every other Nuclide feature, add a Toggle command to toggle terminal visibility. This uses `atom.workspace.toggle` along with the simplified standard uri for terminal implemented prior in this stack.

To align with the behavior of these other features in tool-bar, change the behavior of the terminal button to toggle the terminal. To support the common case of needing another new terminal, use tool-bar's api for handling when alt is pressed while clicking the button.

Reviewed By: rholinshead

Differential Revision: D8878280

fbshipit-source-id: 3a297c101f64b21a514b3cfe476653d523bc0b2b