Debugger Process Tree: add filter to paused threads only button
Summary: This adds a quick filter to show only stopped threads in the debugger process tree. This is especially useful for targets like HHVM where only paused threads are interesting to the user. When set, this button hides all running threads. The state of this button is persisted in featureConfig so it survives between debug sessions and Atom restarts.

Reviewed By: velocityboy

Differential Revision: D9307441

fbshipit-source-id: 51f1752633ab596f1f1513f3b7109b101028d849