fb-intl-fbt-str Rx 2: get translation info using Rx Observable rather than having asnyc code in UI
Summary:
Before we were using an async function call in the `FbtStringBundleRoot` UI component. This is a bit of an anti-pattern. Thus we fix this by moving the `translationInfo` into a part of the `FbtStringBundleRoot`'s properties, and then we bind this component using an `Observable`. This `Observable` emits events in the following cases:
1) When we query the FbtBundle cache
2) When we query the translation info cache

We also modify `getOrCreateFbtBundleTranslationInfo` such that it gets translation information for the full FbtBundle (rather than the bundle passed as a parameter), so that when we update the observables for a given file path (we'll have an observable for each selection-extraction for the file), it updates for the full mapping of hashes to translation info.

Reviewed By: kayhadrin

Differential Revision: D9308005

fbshipit-source-id: 459fb6218f35bb3093ebea1d3762d564c42a8e9c