Don't close tooltips when the user hovers resize handles
Summary:
Since resize handles are invisible, fixed position divs with relatively high z-index, they would trigger mouseleaves on datatips when entered, closing the datatip.

The handles have the same z-index as the dock toggle buttons, which must be greater than the z-index of `scrollbar-content` and inactive `atom-dock`s.

To make naming clearer, this also renames `datatip-overlay` to `datatip-element`, as it actually wasn't added to the `<atom-overlay>` for datatips, but its child. It then uses the `datatip-overlay` name along with a `datatip-pinned-overlay` name to increase the z-index of these overlays. The z-index property is repeated in each as the atom api only allows one class to be added to the overlay element (internally it calls ClassList.prototype.add), though the selectors are combined into one using LESS's `extend`

Reviewed By: matthewwithanm

Differential Revision: D9369101

fbshipit-source-id: 68ccfad88376f56e94698572692aa19c3a5ba220