Debugger process tree: Show warning if the focused thread is hidden by a filter
Summary: If the user has a thread filter that hides the currently focused thread, this creates a confusing experience because if we stop at a breakpoint, the callstack is not visible. Add a warning to the UI to remind the user that he or she has filtered the visible threads, and their filter is excluding the currently focused thread.

Reviewed By: velocityboy

Differential Revision: D9465570

fbshipit-source-id: 3968e189a032aa6121534f0660f3fa340bbc6a8b