Get rid of DisplayableRecord
Summary: After the previous diff in this stack, the only thing that DisplayableRecord was still used for was the expansionStateId. They were always 1:1 with record instances, so we can move then all the way down to the OutputTable and derive them from a WeakMap stored on the instance to ensure they are still unique to this particular console.

Reviewed By: jgebhardt

Differential Revision: D9730154

fbshipit-source-id: e3392dcea97e52db8ed7a303e3a02e700d1a58de