Add lint for unused subscribe result
Summary:
Add an unused-subscription lint which emits a warning when it sees a `.subscribe` at the end of a statement (meaning its return value is unused).

This pattern can cause memory leaks because the closure in the subscription function body holds a reference to the entire context until the underlying observable completes or errors.

In the case of long-running underlying observables, we want to make sure to unsubscribe if we no longer need the result.

Reviewed By: kcaze

Differential Revision: D9949411

fbshipit-source-id: 65acb85349a64b00aca82c628b1362b8a1b84d9d