Nuclide debugger: Remove endlines and endcolumns - 1/N
Summary: In the UX layer we only ever use a single line to represent the location of a breakpoint. Having logic all over the place to decide if various parts of the UI should key off of line or endLine is messy and bug prone. Simplifying to remove this and decide in exactly one place when the protocol breakpoints come back from the debug adapter.

Reviewed By: mostafaeweda

Differential Revision: D9804004

fbshipit-source-id: 8e75a5d49f2180190c238fa766b6e14599c8edb7