Support Multi-Target Breakpoints
Summary:
Currently the UI doesn't properly support breakpoints in multi-target debugging because it only has a single set of breakpoints shared by all adapters. This doesn't work when different adapters have a different notion of which breakpoints are resolved, and what calibrated lines they resolved to.

This diff separates the logic to maintain UI breakpoints (which correspond to locations bps are set in the editor by the user) and adapter breakpoints which have additional info like verified, resolved location, etc.

The UI shows the breakpoints from the currently focused process any time there is one, and falls back to the UI breakpoints otherwise.

Reviewed By: mostafaeweda

Differential Revision: D9803997

fbshipit-source-id: fc131e7914c36d4e4a3d2457c955187b9c611d6a