Click on line number to open file in Commit Info View
Summary:
In the commit info view changed files list, allow clicking on a hunk's line number to open the file at that line number. This makes the list of files extra useful, especially when the commit info view is used as a panel.

This is not super discoverable, since the line numbers don't get underlined when you hover, but that's existing behavior in the AtomTextEditor...

Reviewed By: tjfryan

Differential Revision: D10137825

fbshipit-source-id: 0d05b55697c1d4bd8dd1acdd2da3a6bb581d2173