Reset all users back to "auto" (default canvas) renderer
Summary:
A significant number of users have triggered the slow-render detection and have fallen back to the DOM renderer unnecessisarily. Now that we check if hardware acceleration is enabled before offering to fall back to the DOM renderer, we'll only offer this (generally slower) option in situations where it's really necessary.

This implements the change by changing the configuration key for the renderer. An alternate implementation would be to change the value for the existing key in `setDefaults`, and make sure it's only done once.

Reviewed By: pelmers

Differential Revision: D10177741

fbshipit-source-id: 9da8e7b761bee3e69182da6f64c11db923c6b61b