Add location indicators in the scrollbar
Summary:
This diff adds colorful indicators in the scrollbar to help keep people oriented within a file.

Manual testing shows that performance is quite good. I tested adding 10,000 diagnostics, and it rendered them in about ~10ms (about the same as the time it took to construct the Set containing the indicators). 1,000 diagnostics render in ~5ms.

There are a few additional things I would like to add, but did not get around to:

* Make current cursor position a bit transparent
* Only draw indicators on active editors
* Draw the scroll view on top of the indicators
* Account for folds
* Only show search results when the search panel is open

Reviewed By: wbinnssmith

Differential Revision: D10133506

fbshipit-source-id: f0a564ba4e9aba430cc390fd8199973accb69bc9