Limit diagnostics per provider per file to 1000 and sort them
Summary:
This formally limits the number of diagnostics we store for a given diagnostics provider per file to 1000. The new VSCode extension host-based services limit their own diagnostics to 1000, as does Flow at the LSP layer.

This also sorts these diagnostics as they come in, as the diagnostic message iterator relies on sorted diagnostics per provider to iterate messages in order.

Reviewed By: matthewwithanm

Differential Revision: D10245655

fbshipit-source-id: 7292688181a6d3fda3deafe6e02e9ca76799d642