Don't emit uiConfig changes when there aren't any
Summary:
This stream would frequently emit the same empty array over and over again because while the selector memoized the result, other changes in the store would cause this to `next` unnecessarily.

A comparison of both array's `length` also prevents the common case when providers change, a *new* empty array is emitted.

This is not strictly necessary for the performance improvement for table renders, as the pure component behavior in DiagnosticsView takes care of most of that. This just ensures the stream behaves as it should, and the length comparison saves us a render or two.

Reviewed By: captbaritone

Differential Revision: D10325410

fbshipit-source-id: 3392b5e9bbcbe75c7229241d3f3a3d44b4e7f2ad