Retain references to active subscriptions
Summary:
This was preventing the PerformanceObserver introduced in from collecting data once the observer went out of scope.

A bug in Chrome < 62 frees PerformanceObservers and their listeners
once the PerformanceObserver is eligible for GC:
https://bugs.chromium.org/p/chromium/issues/detail?id=742530

This means that callbacks for a PerformanceObserver will stop getting called
(and themselves GCed) at an abitrary time, unlike other disposables.

Intentionally hold references to all DOM Observers in this set, and delete
them when the last subscriber unsubscribes.

Reviewed By: matthewwithanm

Differential Revision: D10412685

fbshipit-source-id: e9a02396e1d38ee58d01a327bb217d552eee1989