Dispose editor destroy subscriptions when new diagnostics requests come in
Summary:
Previously, every time diagnostics are applied, a subscription to the relevant editor's destruction was created. These subscriptions were never cleaned up until the editor itself was destroyed. However, once the next request creates a subscription, the prior one is no longer necessary.

This also keeps abort controllers around even in the case that applying diagnostics completes, so that when the next request comes in, the prior one is "aborted" even if it has logically completed. This is to keep the logic for cleanup identical in both cases.

Reviewed By: matthewwithanm

Differential Revision: D10387626

fbshipit-source-id: be610d853705fd99be9f347864b76310168e7f11