Use screen positions to compute scrollbar indicator positions
Summary:
Allows scrollbar indicators to respect wordwrap and folding. If a indicator is positioned inside a fold, it is represented as a single line at fold line.

What are the perf implications of all those calls to screenPositionForBufferPosition

Here's a profile of drawing ~600 indicator in ~4.17ms:

{F141648993}

Is there anyway to get notified when a text editor's folds change? Currently it's possible to end up in a situation where some code has folded, but the scrollbar has not yet updated.

Reviewed By: matthewwithanm

Differential Revision: D10452069

fbshipit-source-id: bc56466a59b8bdeda4c53e69fecd4e800b620e0b