Back out "Debugger: Open Dedicated Console Tab for Debugger"
Summary:
This diff, D10228075, introduced a bug which causes the Console to get split multiple times resulting in many more Console pane items than the user expected.

The solution to this is a WIP but in the meanwhile I would like to back out the offending changes so that the next release of Nuclide does not have this bug.

Reviewed By: ebluestein

Differential Revision: D12929905

fbshipit-source-id: b04a10c3c76d2003080f7bacf0135a36dd9515d0