Add a utility to opt out of stalls for user-initiated loading
Summary:
Often times users take an action and can resonably expect a long, blocking task
to run to completion before they can take action again:
https://developers.google.com/web/fundamentals/performance/rail#ux
This is analagous to a web page or app's initial loading, transitioning to
another significant view, etc.

This is a decorator that wraps a function that pauses in response to user action,
opting it out of stall observation and forwarding any arguments passed and returning
the original function's return value.

**Use this cautiously and deliberately, only in situations where it is
reasonable for a user to expect a pause!**

If the action takes longer than 1s, we still record this as a stall, as it
fails the RAIL model's definition of responsive loading.

Reviewed By: aadisriram

Differential Revision: D12893198

fbshipit-source-id: 41b62319ef13244f4ebca19cc2a7d067e01754bc